reworked config directory
