--- CoworkingSpaceReservation (Spring)/src/main/java/org/example/config/AuthInterceptor.java
@@ -1,18 +1,0 @@
-package org.example.config;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-import org.springframework.web.servlet.HandlerInterceptor;
-class AuthInterceptor implements HandlerInterceptor {
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        HttpSession session = request.getSession(false);
-        if (session == null || session.getAttribute("currentUser") == null) {
-            response.sendRedirect("/login");
-            return false;
-        }
-        return true;
-    }
-}
-
