--- CoworkingSpaceReservation (Spring)/src/main/java/org/example/config/DispatcherServletInitializer.java
@@ -1,27 +1,0 @@
-package org.example.config;
-
-import jakarta.servlet.FilterRegistration;
-import jakarta.servlet.ServletContext;
-import jakarta.servlet.ServletRegistration;
-import org.springframework.web.WebApplicationInitializer;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
-import org.springframework.web.filter.HiddenHttpMethodFilter;
-import org.springframework.web.servlet.DispatcherServlet;
-
-public class DispatcherServletInitializer implements WebApplicationInitializer {
-    @Override
-    public void onStartup(ServletContext servletContext){
-        AnnotationConfigWebApplicationContext context = new AnnotationConfigWebApplicationContext();
-        context.register(WebMvcConfiguration.class);
-
-        DispatcherServlet dispatcherServlet = new DispatcherServlet(context);
-
-        ServletRegistration.Dynamic registration = servletContext.addServlet("dispatcher", dispatcherServlet);
-        registration.setLoadOnStartup(1);
-        registration.addMapping("/");
-
-        FilterRegistration.Dynamic filter = servletContext.addFilter("hiddenHttpMethodFilter", new HiddenHttpMethodFilter());
-        filter.addMappingForUrlPatterns(null, false, "/*");
-    }
-}
-
