+++ CoworkingSpaceReservation (Spring)/src/main/java/org/example/config/JpaConfiguration.java
@@ -1,0 +1,9 @@
+package org.example.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@EnableJpaRepositories(basePackages = "org.example.repository")
+public class JpaConfiguration {
+}
