+++ CoworkingSpaceReservation (Spring)/src/main/java/org/example/config/WebMvcConfiguration.java
@@ -1,60 +1,32 @@
 package org.example.config;
 
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.thymeleaf.spring6.SpringTemplateEngine;
-import org.thymeleaf.spring6.templateresolver.SpringResourceTemplateResolver;
-import org.thymeleaf.spring6.view.ThymeleafViewResolver;
-
-import java.nio.charset.StandardCharsets;
 
 @Configuration
-@EnableWebMvc
-@ComponentScan(basePackages = "org.example")
 public class WebMvcConfiguration implements WebMvcConfigurer {
-
-    @Bean
-    public SpringResourceTemplateResolver templateResolver() {
-        SpringResourceTemplateResolver resolver = new SpringResourceTemplateResolver();
-        resolver.setPrefix("/WEB-INF/views/");
-        resolver.setSuffix(".html");
-        resolver.setTemplateMode("HTML");
-        resolver.setCharacterEncoding(StandardCharsets.UTF_8.name());
-        return resolver;
-    }
-
-    @Bean
-    public SpringTemplateEngine templateEngine(SpringResourceTemplateResolver resolver) {
-        SpringTemplateEngine engine = new SpringTemplateEngine();
-        engine.setEnableSpringELCompiler(true);
-        engine.setTemplateResolver(resolver);
-        return engine;
-    }
-
-    @Bean
-    public ThymeleafViewResolver thymeleafViewResolver(SpringTemplateEngine engine) {
-        ThymeleafViewResolver resolver = new ThymeleafViewResolver();
-        resolver.setTemplateEngine(engine);
-        resolver.setOrder(1);
-        resolver.setCharacterEncoding(StandardCharsets.UTF_8.name());
-        return resolver;
-    }
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/css/**")
-                .addResourceLocations("/resources/css/");
-    }
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new AuthInterceptor())
-                .addPathPatterns("/dashboard", "/user/reservations", "/admin/reservations", "user/spaces", "/admin/spaces",
+        InterceptorRegistration interceptor = registry.addInterceptor(new HandlerInterceptor() {
+            @Override
+            public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+                HttpSession session = request.getSession(false);
+                if (session == null || session.getAttribute("currentUser") == null) {
+                    response.sendRedirect("/login");
+                    return false;
+                }
+                return true;
+            }
+        });
+
+        interceptor.addPathPatterns("/dashboard", "/user/reservations", "/admin/reservations", "user/spaces", "/admin/spaces",
                         "/admin/reservations/all", "/user/reservations/list", "/user/reservations/add-reservation-form", "/user/reservations/remove-form",
                         "/admin/spaces/add-spaces-form", "/admin/spaces/remove-spaces-form", "/user/spaces/available", "/user/spaces/available-list")
                 .excludePathPatterns("/login");
